features: fix activity retry
