+++ internal/clusterfeature/clusterfeatureadapter/workflow/common.go
@@ -1,0 +1,31 @@
+// Copyright © 2019 Banzai Cloud
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package workflow
+
+import "errors"
+
+const (
+	shouldRetryReason = "should retry activity"
+)
+
+func shouldRetry(err error) bool {
+	var sh interface {
+		ShouldRetry() bool
+	}
+	if errors.As(err, &sh) {
+		return sh.ShouldRetry()
+	}
+	return false
+}
